refactor: cell-row 연관관계 매핑
Cell Load시 Row도 같이 로드하기위해, 연관관계 매핑 추가
