+++ src/main/java/com/finnect/crm/application/service/cell/CreateNewColumnService.java
@@ -2,7 +2,7 @@ package com.finnect.crm.application.service.cell;
 
 import com.finnect.crm.application.port.in.cell.CreateNewColumnUseCase;
 import com.finnect.crm.application.port.out.cell.SaveCellPort;
-import com.finnect.crm.application.port.out.cell.SaveDataColumnPort;
+import com.finnect.crm.application.port.out.column.SaveDataColumnPort;
 import com.finnect.crm.domain.cell.DataColumn;
 import com.finnect.crm.domain.cell.state.DataColumnState;
 import lombok.RequiredArgsConstructor;
