+++ src/main/java/com/finnect/crm/domain/cell/CellId.java
@@ -1,0 +1,13 @@
+package com.finnect.crm.domain.cell;
+
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@AllArgsConstructor
+@Getter
+public class CellId {
+    private Long columnId;
+    private Long rowId;
+
+}
